Revert

+++ ci_scripts/ci_post_xcodebuild.sh
@@ -5,6 +5,9 @@
 #
 #  Created by Michael Rack on 2/19/25.
 #
+  
+if [ "$CI_XCODEBUILD_ACTION" = "build-for-testing" ]
+then
     rm -rf $CI_RESULT_BUNDLE_PATH
         
     xcodebuild \
@@ -16,15 +19,20 @@
       test-without-building
       
     brew install sonar-scanner
+    
     bash xccov-to-sonarqube-generic.sh /Volumes/workspace/*.xcresult > AAAAA.xml
     cat AAAAA.xml
+    
     echo "==> Full repository dump:"
     find /Volumes/workspace/repository -print
+    
     echo "==> Tree dump of repo (all files):"
     find /Volumes/workspace/repository | sort
+    
     echo "==> Listing contents under expected test paths:"
     find /Volumes/workspace/repository/SQDemoTests -type f
     find /Volumes/workspace/repository/SQDemoUITests -type f
+    
     sonar-scanner \
       -Dsonar.organization=benpatterson48 \
       -Dsonar.projectKey=benpatterson48_SQDemo \
@@ -32,3 +40,6 @@
       -Dsonar.host.url=https://sonarcloud.io \
       -Dsonar.coverageReportPaths=AAAAA.xml \
       -Dsonar.scm.disabled=true
+else
+    echo "==> Not running sonar-scanner steps"
+fi
